edit entity

+++ src/core/entity/post/order.entity.ts
@@ -4,16 +4,16 @@ import { Column, Entity } from "typeorm";
 @Entity('order')
 export class OrderEntity extends BaseEntity{
 
-      @Column({ type: 'bigint'})
+      @Column({ type: 'decimal'})
       total_price: number;
 
-      @Column({ type: 'bigint'})
+      @Column({ type: 'int'})
       quantity: number;
 
-      @Column({ type: 'bigint'})
+      @Column({ type: 'int'})
       customer_id: number;
 
-      @Column({ type: 'bigint'})
+      @Column({ type: 'int'})
       ticket_id: number;
 
       @Column({ type: 'boolean'})
